requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
altgraph==0.17.3
amocrm-api==2.6.1
amocrm-api-wrapper==0.0.17
anyio==3.7.0
appdirs==1.4.4
APScheduler==3.6.3
ascii-magic==1.6
asttokens==2.4.0
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
auto-py-to-exe==2.36.0
Automat==22.10.0
Babel==2.9.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.1.0
blis==0.7.10
boto3==1.34.15
botocore==1.34.15
bottle==0.12.25
bottle-websocket==0.2.9
breadability==0.1.20
bs4==0.0.1
cachetools==4.2.2
catalogue==2.0.9
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
chromedriver-autoinstaller==0.4.0
ci-info==0.3.0
click==8.1.3
click-shell==2.1
cloudscraper==1.2.71
coala-utils==0.5.1
colorama==0.4.6
confection==0.1.3
configobj==5.0.8
configparser==6.0.0
constantly==15.1.0
contourpy==1.1.1
cryptography==41.0.1
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.7
dateutils==0.6.12
decorator==5.1.1
Deprecated==1.2.14
distro==1.8.0
docopt==0.6.2
easygui==0.98.3
Eel==0.16.0
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.1.1
executing==1.2.0
fake-useragent==1.2.1
fastapi==0.92.0
fastapi-queue==0.1.1
filelock==3.12.3
fitz==0.0.1.dev2
fonttools==4.43.1
forex-python==1.8
frozenlist==1.3.3
future==0.18.3
gensim==4.3.2
gevent==22.10.2
gevent-websocket==0.10.1
google-api-core==2.11.0
google-api-python-client==2.104.0
google-auth==2.19.1
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
gspread==5.12.0
gTTS==2.3.2
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.12.0
incremental==22.10.0
ipdb==0.13.13
ipython==8.15.0
isodate==0.6.1
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
langcodes==3.3.0
limits==3.5.0
loader==2017.9.11
loguru==0.7.0
looseversion==1.3.0
lxml==4.9.2
magic-filter==1.0.11
mail-parser==3.15.0
markdown-it-py==2.2.0
markdownify==0.11.6
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
MechanicalSoup==1.3.0
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
mysql-connector==2.2.9
mysql-connector-python==8.1.0
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
nltk==3.8.1
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.2.2
opencv-python==4.8.1.78
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
panda==0.3.1
pandas==2.0.2
paramiko==3.2.0
parse==1.19.1
parsel==1.8.1
parso==0.8.3
pathlib==1.0.1
pathy==0.10.2
patool==1.15.0
pdfminer.six==20221105
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.5.0
preshed==3.0.8
prompt-toolkit==3.0.39
Protego==0.3.0
protobuf==4.21.12
prov==2.0.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==23.12.11
pycparser==2.21
pydantic==1.10.8
PyDispatcher==2.0.7
pydot==2.0.0
pydub==0.25.1
pyee==8.2.2
Pygments==2.15.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.3
PyJWT==2.7.0
PyMuPDF==1.23.9
PyMuPDFb==1.23.9
PyMySQL==1.1.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
pyTelegramBotAPI==4.12.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.1
python-multipart==0.0.6
python-telegram-bot==20.7
pytz==2023.3
PyVirtualDisplay==3.0
pywin32-ctypes==0.2.1
pyxnat==1.6
PyYAML==6.0
queuelib==1.6.2
rarfile==4.1
ratelimit==2.2.1
rdflib==7.0.0
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
requests-html==0.10.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.4.1
rsa==4.9
s3transfer==0.10.0
schedule==1.2.1
scipy==1.11.3
Scrapy==2.10.1
scrapy-selenium==0.0.7
seaborn==0.13.0
selenium==4.9.1
service-identity==23.1.0
shellingham==1.5.0.post1
simplejson==3.19.1
six==1.16.0
slowapi==0.1.8
smart-open==6.4.0
sniffio==1.3.0
snscrape==0.7.0.20230622
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.10.0
sqlite3-api==2.0.4
srsly==2.4.7
sshtunnel==0.4.0
stack-data==0.6.2
starlette==0.25.0
sumy==0.11.0
tabula==1.0.5
tabula-py==2.9.0
tabulate==0.9.0
telebot==0.0.5
thinc==8.1.12
ThreadPoolExecutorPlus==0.2.2
tldextract==3.5.0
tomli==2.0.1
tornado==6.4
tqdm==4.65.0
traitlets==5.9.0
traits==6.3.2
trio==0.22.0
trio-websocket==0.10.2
tweety-ns==0.9.9.1
Twisted==22.10.0
twisted-iocpsupport==1.0.4
twitter-scraper==0.4.4
twitterscraper==1.6.1
typer==0.4.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
undetected-chromedriver==3.4.7
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.20.0
w3lib==2.1.2
wasabi==1.1.2
wcwidth==0.2.6
websockets==10.4
whichcraft==0.6.1
win32-setctime==1.1.0
wrapt==1.15.0
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
youdotcom==2.0.25
zipp==3.15.0
zope.event==5.0
zope.interface==6.0
